requirements.txt
# Core dependencies
gradio>=4.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Quality assessment
language-tool-python>=2.7.0
textstat>=0.7.3

# Configuration
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0


